requirements.txt
python-dotenv==1.0.0
questionary==2.0.1
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
requests==2.31.0
pydantic==2.4.2
